requirements.txt
annotated-types==0.7.0
anyio==4.9.0
blinker==1.9.0
certifi==2025.4.26
cffi==1.17.1
click==8.2.1
distro==1.9.0
exceptiongroup==1.3.0
Flask==3.1.1
flatlib==0.2.3
geographiclib==2.0
geopy==2.4.1
gunicorn==23.0.0
h11==0.16.0
h3==4.2.2
httpcore==1.0.9
httpx==0.28.1
idna==3.10
itsdangerous==2.2.0
Jinja2==3.1.6
jiter==0.10.0
MarkupSafe==3.0.2
numpy==2.2.6
openai==1.82.1
packaging==25.0
pycparser==2.22
pydantic==2.11.5
pydantic_core==2.33.2
pyswisseph==2.8.0.post1
python-dotenv==1.1.0
pytz==2025.2
sniffio==1.3.1
timezonefinder==6.5.9
tqdm==4.67.1
typing-inspection==0.4.1
typing_extensions==4.13.2
Werkzeug==3.1.3
